refactor: cleanURL 을 커스텀 훅으로 분리
