+++ src/components/CardForm/index.tsx
@@ -7,11 +7,10 @@ import {
   useRef,
   useState,
 } from 'react';
-import { useLocation } from 'react-router-dom';
 
 import { CardStep } from '../../constants';
 import CardFormContext from '../../contexts/CardFormContext';
-import useMoveToPage from '../../hooks/useMoveToPage';
+import { useCleanURL, useMoveToPage } from '../../hooks';
 import CardCompanySelect from '../CardCompanySelect';
 import CardCVCInput from '../CardCVCInput';
 import CardExpirationPeriodInput from '../CardExpirationPeriodInput';
@@ -27,7 +26,7 @@ function CardForm() {
   const { navigateToPage } = useMoveToPage('cardConfirmation');
   const [openFormFields, setOpenFormFields] = useState<CardStep[]>(['numbers']);
   const layoutRef = useRef<HTMLElement>();
-  const location = useLocation();
+  useCleanURL();
 
   const goNextFormStep = (currentStep: CardStep) => {
     setOpenFormFields((prev) => [...prev, currentStep]);
@@ -85,21 +84,7 @@ function CardForm() {
     cardFormContext.resetCardInfo();
   };
 
-  /**
-   * url 변경 시 실행될 함수
-   * 카드 등록 페이지로 돌아오거나 새로고침 시 url에 form의 입력 필드의 값들이 포함되는 오류를 수정하기 위해 사용
-   */
-  const cleanUpURL = () => {
-    const { search } = location;
-    // 쿼리 문자열이 있는 경우
-    if (search) {
-      const newURL = window.location.origin + window.location.pathname;
-      window.history.replaceState(null, '', newURL);
-    }
-  };
-
   useEffect(() => {
-    cleanUpURL();
     const $layout = document.getElementById('layout');
     if ($layout) layoutRef.current = $layout;
   }, []);
