+++ src/hooks/index.ts
@@ -1,5 +1,6 @@
 import useCardInput from './useCardInput';
+import useCleanURL from './useCleanURL';
 import useMoveToPage from './useMoveToPage';
 import useNextFormStep from './useNextFormStep';
 
-export { useCardInput, useMoveToPage, useNextFormStep };
+export { useCardInput, useMoveToPage, useNextFormStep, useCleanURL };
